requirements.txt
pip==20.1.1

celery

# Telegram
python-telegram-bot

# Rendering
typus
reportlab
pdf2image

# Sentry
sentry-sdk==0.15.1
